Use abstract class for split tasks
